+++ client/asset-plugin-posix/amapy_plugin_posix/transporter/posix_transport_resource.py
@@ -1,7 +1,7 @@
 from cached_property import cached_property
 
 from amapy_pluggy.storage.transporter import TransportResource
-from asset_plugin_posix.posix_url import PosixURL
+from amapy_plugin_posix.posix_url import PosixURL
 
 
 class PosixTransportResource(TransportResource):
